+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Added an option to make mobs find a better path to the target
+* Pearler Zombies inaccuracy can now be configured
 * Allow Climbing and Target Ladders are now disabled by default
 * Entities with higher follow range than the override are no longer overridden (e.g. Ghasts)
 * Creepers now start exploding when nearer to the target (explosion radius ~~x1.5~~ -> 1.35 blocks)